python/requirements.txt
lz4==4.3.2
kafka-python==2.0.2
fastavro==1.8.3